test: 테스트 코드 각각의 객체로 이동
