+++ src/test/java/domain/DealerTest.java
@@ -1,0 +1,62 @@
+package domain;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class DealerTest {
+
+    @DisplayName("딜러는 처음 받은 2장의 카드의 합이 16 이하라면 카드를 뽑는다")
+    @Test
+    void canHit() {
+        //given
+        Cards cards = new Cards();
+        Card card1 = new Card(Symbol.COLVER, Number.ACE);
+        Card card2 = new Card(Symbol.HEART, Number.TWO);
+        Card card3 = new Card(Symbol.SPADE, Number.KING);
+
+        cards.add(card1);
+        cards.add(card2);
+        cards.add(card3);
+
+        Dealer dealer = new Dealer();
+        dealer.prepareGame(cards);
+
+        int initialPoint = dealer.getCards().calculateTotalPoint();
+
+        //when
+        dealer.hit(cards);
+        int actual = dealer.getCards().calculateTotalPoint();
+
+        //then
+        assertThat(actual).isNotEqualTo(initialPoint);
+    }
+
+    @DisplayName("딜러는 처음 받은 2장의 카드의 합이 17 이상이라면 카드를 뽑지 않는다")
+    @Test
+    void cannotHit() {
+        //given
+        Cards cards = new Cards();
+        Card card1 = new Card(Symbol.COLVER, Number.JACK);
+        Card card2 = new Card(Symbol.HEART, Number.NINE);
+        Card card3 = new Card(Symbol.SPADE, Number.KING);
+
+        cards.add(card1);
+        cards.add(card2);
+        cards.add(card3);
+
+        Dealer dealer = new Dealer();
+        dealer.prepareGame(cards);
+
+        int initialPoint = dealer.getCards().calculateTotalPoint();
+
+        //when
+        dealer.hit(cards);
+        int actual = dealer.getCards().calculateTotalPoint();
+
+        //then
+        assertThat(actual).isEqualTo(initialPoint);
+    }
+
+}
